openai/processor と dify/processor で共通して使用しそうな処理を base/processor に移動
